# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Novex Backend API
  version: 1.1.0
  description: |
    REST API for NovexPanel backend.

    Notes:
    - Auth-protected routes require JWT Bearer token.
    - WebSocket endpoints are documented for handshake and query params.
    - Deploy listing supports server filter and pagination parameters.
servers:
  - url: http://localhost:8380
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ServerIdPath:
      in: path
      name: id
      required: true
      description: Server identifier.
      schema:
        type: integer
        minimum: 1
    DeployIdPath:
      in: path
      name: id
      required: true
      description: Deploy identifier.
      schema:
        type: integer
        minimum: 1
    ProcessPidPath:
      in: path
      name: pid
      required: true
      description: Process PID.
      schema:
        type: integer
        minimum: 1
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: JWT token for API and site websocket.
    AgentTokenCreateRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          example: Production server token
    AgentTokenCreateResponse:
      type: object
      properties:
        id:
          type: integer
        agent_token:
          type: string
          description: Full token shown only once.
        token_prefix:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        name:
          type: string
          nullable: true
    AgentTokenLinkedServer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ip:
          type: string
        online:
          type: boolean
    AgentTokenListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        token_prefix:
          type: string
        revoked:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        server:
          allOf:
            - $ref: '#/components/schemas/AgentTokenLinkedServer'
          nullable: true
    AgentTokenRenameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ServerItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ip:
          type: string
        online:
          type: boolean
        last_metrics:
          type: object
          additionalProperties: true
...
    RunCommandRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
    ProcessItem:
      type: object
      properties:
        pid:
          type: integer
        name:
          type: string
        cpu:
          type: number
        mem:
          type: number
    ProcessListResponse:
      type: object
      properties:
        processes:
          type: array
          items:
            $ref: '#/components/schemas/ProcessItem'
    KillProcessResponse:
      type: object
      properties:
        success:
          type: boolean
    LegacyServerDeployRequest:
      type: object
      required: [source]
      properties:
        source:
          type: string
          enum: [github, zip]
        repo_url:
          type: string
        branch:
          type: string
        project_type:
          type: string
          enum: [auto, node, react, vue, svelte, vite, go, docker, static, python]
        zip_data:
          type: string
          description: Base64-encoded zip archive if source=zip.
    DeployCreateRequest:
      type: object
      required: [serverId, repoUrl]
      properties:
        serverId:
          type: integer
          minimum: 1
        repoUrl:
          type: string
          example: https://github.com/acme/my-frontend.git
        branch:
          type: string
          default: main
        type:
          type: string
          enum: [auto, static, node, vite, react, vue, svelte, go, python]
          default: auto
        subdirectory:
          type: string
          description: Relative path inside repository to deploy from.
          example: frontend
        envVars:
          type: object
          additionalProperties:
            type: string
          description: Optional environment variables passed to runtime container.
        buildCommand:
          type: string
          description: Optional custom build command. If empty, default build strategy is used.
          example: npm ci && npm run build
        outputDir:
          type: string
          description: Explicit static build output directory.
          example: dist
    DeployCreateAcceptedResponse:
      type: object
      properties:
        deployId:
          type: integer
        status:
          type: string
          example: pending
        error:
          type: string
          nullable: true
    DeployItem:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        repoUrl:
          type: string
        branch:
          type: string
        type:
          type: string
        subdirectory:
          type: string
        buildCommand:
          type: string
          nullable: true
        outputDir:
          type: string
          nullable: true
        status:
          type: string
          example: running
        url:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        deployLogPreview:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeployListResponse:
      description: |
        Deploy list response.
        Current backend returns an array. Paginated object form is documented for compatibility with clients that request pagination metadata.
      oneOf:
        - type: array
          items:
            $ref: '#/components/schemas/DeployItem'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/DeployItem'
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
    DeployDetailsResponse:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        repoUrl:
          type: string
        branch:
          type: string
        type:
          type: string
        subdirectory:
          type: string
        buildCommand:
          type: string
          nullable: true
        outputDir:
          type: string
          nullable: true
        status:
          type: string
        url:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeployLogResponse:
      type: object
      properties:
        deployId:
          type: integer
        status:
          type: string
        url:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        type:
          type: string
        subdirectory:
          type: string
        deployLog:
          type: string
    DeployLogItem:
      type: object
      properties:
        id:
          type: integer
        deploy_id:
          type: integer
        line:
          type: string
        is_error:
          type: boolean
        created_at:
          type: string
          format: date-time
    DeployLogsResponse:
      type: object
      properties:
        deploy_id:
          type: integer
        status:
          type: string
        url:
          type: string
          nullable: true
        logs:
          type: array
          items:
            $ref: '#/components/schemas/DeployLogItem'
    DeployDeleteResponse:
      type: object
      properties:
        deployId:
          type: integer
        status:
          type: string
          example: deleted
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                value:
                  email: user@example.com
                  password: secret123
      responses:
        '201':
          description: User created.
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid:
                  value:
                    error: invalid request body
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              basic:
                value:
                  email: user@example.com
                  password: secret123
      responses:
        '200':
          description: JWT issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  value:
                    error: invalid credentials
  /auth/tokens:
    post:
      tags: [Auth]
      summary: Create agent token
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTokenCreateRequest'
      responses:
        '201':
          description: Token created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTokenCreateResponse'
    get:
      tags: [Auth]
      summary: List agent tokens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentTokenListItem'
  /auth/tokens/{id}:
    patch:
      tags: [Auth]
      summary: Rename agent token
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTokenRenameRequest'
      responses:
        '204':
          description: Token renamed.
    delete:
      tags: [Auth]
      summary: Revoke agent token
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Token revoked.
  /servers:
    get:
      tags: [Servers]
      summary: List user servers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Server list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServerItem'
...
  /servers/{id}/processes:
    get:
      tags: [Servers]
      summary: List processes on a server
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerIdPath'
      responses:
        '200':
          description: Process list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessListResponse'
              examples:
                sample:
                  value:
                    processes:
                      - pid: 1052
                        name: node
                        cpu: 12.8
                        mem: 3.4
                      - pid: 1120
                        name: nginx
                        cpu: 1.1
                        mem: 0.8
        '404':
          description: Server not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    error: server not found
  /servers/{id}/processes/{pid}:
    delete:
      tags: [Servers]
      summary: Kill process on a server
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerIdPath'
        - $ref: '#/components/parameters/ProcessPidPath'
      responses:
        '200':
          description: Kill command result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KillProcessResponse'
              examples:
                success:
                  value:
                    success: true
        '400':
          description: Invalid pid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_pid:
                  value:
                    error: invalid pid
  /servers/{id}/command:
    post:
      tags: [Servers]
      summary: Execute shell command on server
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCommandRequest'
      responses:
        '200':
          description: Command result.
        '502':
          description: Agent communication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /servers/{id}/deploy:
    post:
      tags: [Servers]
      summary: Start legacy deploy flow on server
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyServerDeployRequest'
      responses:
        '202':
          description: Deploy accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deploy_id:
                    type: integer
                  status:
                    type: string
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deploy:
    post:
      tags: [Deploys]
      summary: Create deploy
      description: Create and dispatch deploy for selected server.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployCreateRequest'
            examples:
              vite_project:
                value:
                  serverId: 12
                  repoUrl: https://github.com/acme/marketing-site.git
                  branch: main
                  type: vite
                  subdirectory: web
                  envVars:
                    NODE_ENV: production
                  buildCommand: ''
                  outputDir: dist
      responses:
        '202':
          description: Deploy accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployCreateAcceptedResponse'
              examples:
                accepted:
                  value:
                    deployId: 452
                    status: pending
                accepted_with_dispatch_error:
                  value:
                    deployId: 452
                    status: failed
                    error: agent offline
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_server:
                  value:
                    error: serverId is required
        '404':
          description: Server not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                server_not_found:
                  value:
                    error: server not found
  /deploys:
    get:
      tags: [Deploys]
      summary: List deploys
      description: List deploys with server filter and pagination parameters.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: serverId
          required: true
          description: Filter by server id.
          schema:
            type: integer
            minimum: 1
        - in: query
          name: page
          required: false
          description: 1-based page number.
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          required: false
          description: Number of items per page.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - in: query
          name: subdirectory
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Deploy list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployListResponse'
              examples:
                array_form:
                  value:
                    - id: 452
                      serverId: 12
                      repoUrl: https://github.com/acme/marketing-site.git
                      branch: main
                      type: vite
                      subdirectory: web
                      buildCommand: ''
                      outputDir: dist
                      status: running
                      url: http://10.0.0.11:32789
                      port: 32789
                      deployLogPreview: building static assets...
                      createdAt: '2026-04-23T10:00:00Z'
                      updatedAt: '2026-04-23T10:02:10Z'
                paginated_form:
                  value:
                    items: []
                    page: 1
                    pageSize: 20
                    total: 0
        '400':
          description: Invalid query params.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_server:
                  value:
                    error: serverId query param is required
  /deploys/{id}:
    get:
      tags: [Deploys]
      summary: Get deploy by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeployIdPath'
      responses:
        '200':
          description: Deploy details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployDetailsResponse'
              examples:
                success:
                  value:
                    id: 452
                    serverId: 12
                    repoUrl: https://github.com/acme/marketing-site.git
                    branch: main
                    type: vite
                    subdirectory: web
                    buildCommand: ''
                    outputDir: dist
                    status: running
                    url: http://10.0.0.11:32789
                    port: 32789
                    createdAt: '2026-04-23T10:00:00Z'
                    updatedAt: '2026-04-23T10:02:10Z'
        '404':
          description: Deploy not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    error: deploy not found
    delete:
      tags: [Deploys]
      summary: Stop and delete deploy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeployIdPath'
      responses:
        '200':
          description: Deploy deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployDeleteResponse'
              examples:
                deleted:
                  value:
                    deployId: 452
                    status: deleted
        '502':
          description: Failed to notify agent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                agent_error:
                  value:
                    error: agent offline
  /deploys/{id}/log:
    get:
      tags: [Deploys]
      summary: Get aggregated deploy log
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeployIdPath'
      responses:
        '200':
          description: Aggregated deploy log and status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployLogResponse'
              examples:
                success:
                  value:
                    deployId: 452
                    status: running
                    url: http://10.0.0.11:32789
                    port: 32789
                    type: vite
                    subdirectory: web
                    deployLog: |
                      cloning repository...
                      npm install
                      npm run build
        '404':
          description: Deploy not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deploys/{id}/logs:
    get:
      tags: [Deploys]
      summary: Get structured deploy log entries
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeployIdPath'
      responses:
        '200':
          description: Structured deploy logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployLogsResponse'
  /site/ws:
    get:
      tags: [WebSocket]
      summary: Site websocket handshake
      description: Upgrade endpoint for UI websocket channel.
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
          description: JWT from /auth/login.
      responses:
        '101':
          description: Switching Protocols.
        '401':
          description: Invalid token.
  /agent/ws:
    get:
      tags: [WebSocket]
      summary: Agent websocket handshake
      description: Upgrade endpoint for agent channel.
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
          description: Agent token from /auth/tokens.
        - in: query
          name: name
          required: false
          schema:
            type: string
          description: Optional server display name.
      responses:
        '101':
          description: Switching Protocols.
        '401':
          description: Invalid or expired token.